Put together the first stable version of the app
Thinking through the component structure of the website as well as
the state management among them represented the greatest hurdles
of the project.

There was some trouble to set the Webpack up to load images dynamically
within a few instances of the Card component, contexts and new loaders
came to my rescue here. I also had my first experience bringing on a
font from an unknown webite, the official "The Office font" will
contribute to make up a remarkable visual identity.

Only functional components came in play so as to practice using Hooks.